Implements sticky footer for home page (#48)
Fixes https://github.com/kubetail-org/kubetail/issues/46
